Add Kalman filter to morse decoder as a try Move All device related command to CCommand Add a bandscope widget

+++ src/backend/IDemodulator.cpp
@@ -35,7 +35,7 @@ void IDemodulator::slotFrequency(double value)
     return;
 }
 
-void IDemodulator::slotBandwidth(int value)
+void IDemodulator::slotBandwidth(double value)
 {
     return;
 }
